modified

+++ STRING/PanagramChecking.java
@@ -8,8 +8,7 @@ public class PanagramChecking {
     public static void main(String[] args) {
 
         String str="Bawds jog, flick quartz, vex nymph";
-        System.out.println(str.charAt(0));
-        
+        System.out.println((int)str.charAt(0));
     }
 //    public static boolean checkPangram  (String s) {
 //        int [] ALPHABET = new int[26];
